+++ android/expoview/src/main/java/versioned/host/exp/exponent/ExperiencePackagePicker.kt
@@ -47,7 +47,7 @@ import expo.modules.network.NetworkModule
 import expo.modules.notifications.NotificationsPackage
 import expo.modules.permissions.PermissionsPackage
 import expo.modules.random.RandomModule
-import expo.modules.screencapture.ScreenCapturePackage
+import expo.modules.screencapture.ScreenCaptureModule
 import expo.modules.screenorientation.ScreenOrientationModule
 import expo.modules.securestore.SecureStorePackage
 import expo.modules.sensors.SensorsPackage
@@ -86,7 +86,6 @@ object ExperiencePackagePicker : ModulesProvider {
     NotificationsPackage(),
     PermissionsPackage(),
     SQLitePackage(),
-    ScreenCapturePackage(),
     SecureStorePackage(),
     SensorsPackage(),
     SpeechPackage(),
@@ -134,6 +133,7 @@ object ExperiencePackagePicker : ModulesProvider {
     MediaLibraryModule::class.java,
     NetworkModule::class.java,
     RandomModule::class.java,
+    ScreenCaptureModule::class.java,
     ScreenOrientationModule::class.java,
     SMSModule::class.java,
     SharingModule::class.java,