+++ packages/expo-screen-capture/src/ExpoScreenCapture.ts
@@ -1,3 +1,2 @@
-import { NativeModulesProxy } from 'expo-modules-core';
-
-export default NativeModulesProxy.ExpoScreenCapture;
+import { requireNativeModule } from 'expo-modules-core';
+export default requireNativeModule('ExpoScreenCapture');